services/vjepa/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
Pillow>=9.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.28.0
boto3>=1.26.0
scikit-image>=0.21.0
